collision-examples - platformer collision example.
There's actually a fair amount of stuff going on in here. It's also
nearly a game - all we need is a score counter and a system for spawning
more fireballs and we're set.

Some lines need to be un-commented to unlock the full functionality of
the example. During class we'll work through it one step at a time.